requirements.txt
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=9.0.0
scipy>=1.7.0